+++ integrations/clickup/main.py
@@ -1,4 +1,5 @@
 from enum import StrEnum
+from typing import Any, Callable, Coroutine
 
 from loguru import logger
 from port_ocean.context.ocean import ocean
@@ -15,6 +16,18 @@ class ObjectKind(StrEnum):
 
 def get_client() -> ClickupClient:
     return ClickupClient(ocean.integration_config["clickup_apikey"])
+
+
+async def setup_application() -> None:
+    app_host = ocean.integration_config.get("app_host")
+    if not app_host:
+        logger.warning(
+            "No app host provided, skipping webhook creation. "
+            "Without setting up the webhook, the integration will not export live changes from Clickup"
+        )
+        return
+    client = get_client()
+    await client.create_webhook_events(app_host)
 
 
 @ocean.on_resync(ObjectKind.TEAM)
@@ -37,3 +50,53 @@ async def on_resync_issues(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     async for issues in client.get_paginated_issues():
         logger.info(f"Received issue batch with {len(issues)} issues")
         yield issues
+
+
+async def handle_list_created_or_updated(data: dict[str, Any]) -> None:
+    client = get_client()
+    project = await client.get_single_project(data["list_id"])
+    await ocean.register_raw(ObjectKind.PROJECT, [project])
+
+
+async def handle_list_deleted(data: dict[str, Any]) -> None:
+    await ocean.unregister_raw(ObjectKind.PROJECT, [{"id": data["list_id"]}])
+
+
+async def handle_task_created_or_updated(data: dict[str, Any]) -> None:
+    client = get_client()
+    issue = await client.get_single_issue(data["task_id"])
+    await ocean.register_raw(ObjectKind.ISSUE, [issue])
+
+
+async def handle_task_deleted(data: dict[str, Any]) -> None:
+    await ocean.unregister_raw(ObjectKind.ISSUE, [{"id": data["task_id"]}])
+
+
+@ocean.router.post("/webhook")
+async def handle_webhook_request(data: dict[str, Any]) -> dict[str, Any]:
+    event_handlers: dict[str, Callable[[dict[str, Any]], Coroutine[Any, Any, None]]] = {
+        "listCreated": handle_list_created_or_updated,
+        "listUpdated": handle_list_created_or_updated,
+        "listDeleted": handle_list_deleted,
+        "taskCreated": handle_task_created_or_updated,
+        "taskUpdated": handle_task_created_or_updated,
+        "taskDeleted": handle_task_deleted,
+    }
+    event: str = data.get("event", "")
+    logger.info(f"Received webhook event of type: {event}")
+    handler = event_handlers.get(event, None)
+    if not handler:
+        logger.error(f"Unhandled event type: {event}")
+        return {"ok": False}
+    await handler(data)
+    logger.info("Webhook event processed")
+    return {"ok": True}
+
+
+@ocean.on_start()
+async def on_start() -> None:
+    logger.info("Starting Port Ocean Clickup integration")
+    if ocean.event_listener_type == "ONCE":
+        logger.info("Skipping webhook creation because the event listener is ONCE")
+        return
+    await setup_application()