+++ integrations/clickup/CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
-# Port_Ocean 0.0.1 (2024-07-04)
+# Port_Ocean 0.1.0 (2024-07-04)
 
 ### Features
 
-- Implemented Clickup integration to ingest teams, projects and issues (0.0.1)
+- Implemented Clickup integration to ingest teams, projects and issues (0.1.0)